+++ addons/curved_lines_2d/assign_collision_inspector_form.gd
@@ -19,7 +19,8 @@ func _enter_tree() -> void:
 	create_button = find_child("CreateCollisionButton")
 	select_button = find_child("GotoCollisionButton")
 	title_button = find_child("TitleButton")
-	scalable_vector_shape_2d.assigned_node_changed.connect(_on_svs_assignment_changed)
+	if 'assigned_node_changed' in scalable_vector_shape_2d:
+		scalable_vector_shape_2d.assigned_node_changed.connect(_on_svs_assignment_changed)
 	collapsible_siblings = get_children().filter(func(x): return x != title_button and not x is Label)
 	_on_svs_assignment_changed()