+++ addons/curved_lines_2d/assign_fill_inspector_form.gd
@@ -1,5 +1,5 @@
 @tool
-extends Control
+extends KeyframeButtonCapableInspectorFormBase
 
 class_name AssignFillInspectorForm
 
@@ -30,9 +30,16 @@ func _enter_tree() -> void:
 	linear_gradient_toggle_button = find_child("LinearGradientToggleButton")
 	radial_gradient_toggle_button = find_child("RadialGradientToggleButton")
 	other_texture_toggle_button = find_child("OtherTextureToggleButton")
-	scalable_vector_shape_2d.assigned_node_changed.connect(_on_svs_assignment_changed)
+	if 'assigned_node_changed' in scalable_vector_shape_2d:
+		scalable_vector_shape_2d.assigned_node_changed.connect(_on_svs_assignment_changed)
 	collapsible_siblings = get_children().filter(func(x): return x != title_button and not x is Label)
 	_on_svs_assignment_changed()
+	_initialize_keyframe_capabilities()
+
+
+func _on_key_frame_capabilities_changed():
+	find_child("AddFillKeyFrameButton").visible = _is_key_frame_capable()
+	find_child("BatchInsertGradientKeyFrameButton").visible = _is_key_frame_capable()
 
 
 func _on_svs_assignment_changed() -> void:
@@ -217,3 +224,40 @@ static func _initialize_gradient(box : Rect2) -> GradientTexture2D:
 	texture.gradient.colors = [Color.WHITE, Color.BLACK]
 	texture.gradient.offsets = [0.0, 1.0]
 	return texture
+
+
+func _on_add_fill_key_frame_button_pressed() -> void:
+	if is_instance_valid(scalable_vector_shape_2d.polygon):
+		add_key_frame(
+			scalable_vector_shape_2d.polygon, "color", color_button.color
+		)
+
+
+func _on_batch_insert_gradient_key_frame_button_pressed() -> void:
+	if not is_instance_valid(scalable_vector_shape_2d.polygon):
+		return
+	if not scalable_vector_shape_2d.polygon.texture is GradientTexture2D:
+		return
+
+	var p2d := scalable_vector_shape_2d.polygon
+	var animation_name := animation_under_edit_button.get_item_text(animation_under_edit_button.get_selected_id())
+	var animation_player := _find_animation_player(animation_name)
+	var track_position := _guarded_get_track_position()
+	var animation := _guarded_get_animation(animation_player)
+	var path_to_node := _guarded_get_path_to_node(animation_player, p2d)
+	if not animation:
+		return
+	if path_to_node.is_empty():
+		return
+	var undo_redo := EditorInterface.get_editor_undo_redo()
+	undo_redo.create_action("Batch all gradient keyframes for %s on animation %s" % [str(p2d), str(animation)])
+	_add_key_frame(undo_redo, animation, NodePath("%s:texture:gradient:colors" % path_to_node),
+		track_position, p2d.texture.gradient.colors)
+	_add_key_frame(undo_redo, animation, NodePath("%s:texture:gradient:offsets" % path_to_node),
+		track_position, p2d.texture.gradient.offsets)
+	_add_key_frame(undo_redo, animation, NodePath("%s:texture:fill_from" % path_to_node),
+		track_position, p2d.texture.fill_from)
+	_add_key_frame(undo_redo, animation, NodePath("%s:texture:fill_to" % path_to_node),
+		track_position, p2d.texture.fill_to)
+
+	undo_redo.commit_action()